Update MonitoringSetUp.sh
Spacing changes

+++ MonitoringSetUp.sh
@@ -45,28 +45,28 @@ else
    GCS_CERT=$GCS_CERT_FOLDER/gcscert.pem
    GCS_KEY=$GCS_CERT_FOLDER/gcskey.pem
 
-	if [ -f "$GCS_CERT" ]; then
-		sudo rm -f "$GCS_CERT"
-	fi
+   if [ -f "$GCS_CERT" ]; then
+      sudo rm -f "$GCS_CERT"
+   fi
 
-	if [ -f "$GCS_KEY" ]; then
-		sudo rm -f "$GCS_KEY"
-	fi
+   if [ -f "$GCS_KEY" ]; then
+      sudo rm -f "$GCS_KEY"
+   fi
 
-    if [ -f "$GCS_CERT_WITH_KEY" ]; then
-	  sudo openssl x509 -in "$GCS_CERT_WITH_KEY" -out "$GCS_CERT"  && sudo chmod 744 "$GCS_CERT" 
-	  sudo openssl pkey -in "$GCS_CERT_WITH_KEY" -out "$GCS_KEY" && sudo chmod 744 "$GCS_KEY" 
-	  sudo rm -f "$GCS_CERT_WITH_KEY" 
-    fi
-	
-	## Create Environment variable files for MDS and MDM
-	echo -e "\n\n###################################### Creating Environment variable files for MDS and MDM #####################\n\n"
+   if [ -f "$GCS_CERT_WITH_KEY" ]; then
+      sudo openssl x509 -in "$GCS_CERT_WITH_KEY" -out "$GCS_CERT"  && sudo chmod 744 "$GCS_CERT" 
+      sudo openssl pkey -in "$GCS_CERT_WITH_KEY" -out "$GCS_KEY" && sudo chmod 744 "$GCS_KEY" 
+      sudo rm -f "$GCS_CERT_WITH_KEY" 
+   fi
+    
+    ## Create Environment variable files for MDS and MDM
+    echo -e "\n\n###################################### Creating Environment variable files for MDS and MDM #####################\n\n"
 
 
 sudo rm -f /tmp/collectd
 cat > /tmp/collectd <<EOT
 # Setting Environment variables for Monitoring
-		 
+         
 export MONITORING_TENANT=$tenant
 export MONITORING_ROLE=GHPI
 export MONITORING_ROLE_INSTANCE=${tenant}_1
@@ -78,36 +78,36 @@ MDSD_OPTIONS="-A -c /etc/mdsd.d/mdsd.xml -d -r $MDSD_ROLE_PREFIX -e $MDSDLOG/mds
 
 sudo rm -f /tmp/mdsd
 cat > /tmp/mdsd <<EOT
-	# Check 'mdsd -h' for details.
+    # Check 'mdsd -h' for details.
 
-	# MDSD_OPTIONS="-d -r ${MDSD_ROLE_PREFIX}"
+    # MDSD_OPTIONS="-d -r ${MDSD_ROLE_PREFIX}"
 
-	MDSD_OPTIONS="-A -c /etc/mdsd.d/mdsd.xml -d -r $MDSD_ROLE_PREFIX -e $MDSDLOG/mdsd.err -w $MDSDLOG/mdsd.warn -o $MDSDLOG/mdsd.info"
+    MDSD_OPTIONS="-A -c /etc/mdsd.d/mdsd.xml -d -r $MDSD_ROLE_PREFIX -e $MDSDLOG/mdsd.err -w $MDSDLOG/mdsd.warn -o $MDSDLOG/mdsd.info"
 
-	export MONITORING_GCS_ENVIRONMENT=Test
+    export MONITORING_GCS_ENVIRONMENT=Test
 
-	export MONITORING_GCS_ACCOUNT=GHPILOGS
+    export MONITORING_GCS_ACCOUNT=GHPILOGS
 
-	export MONITORING_GCS_REGION=westus
-	# or, pulling data from IMDS
+    export MONITORING_GCS_REGION=westus
+    # or, pulling data from IMDS
 
-	# imdsURL="http://169.254.169.254/metadata/instance/compute/location?api-version=2017-04-02&format=text"
+    # imdsURL="http://169.254.169.254/metadata/instance/compute/location?api-version=2017-04-02&format=text"
 
-	# export MONITORING_GCS_REGION="$(curl -H Metadata:True --silent $imdsURL)"
+    # export MONITORING_GCS_REGION="$(curl -H Metadata:True --silent $imdsURL)"
 
-	# see https://jarvis.dc.ad.msft.net/?section=b7a73824-bbbf-49fc-8c3e-a97c27a7659e&page=documents&id=66b7e29f-ddd6-4ab9-ad0a-dcd3c2561090
+    # see https://jarvis.dc.ad.msft.net/?section=b7a73824-bbbf-49fc-8c3e-a97c27a7659e&page=documents&id=66b7e29f-ddd6-4ab9-ad0a-dcd3c2561090
 
-	export MONITORING_GCS_CERT_CERTFILE="$GCS_CERT"   # update for your cert on disk
+    export MONITORING_GCS_CERT_CERTFILE="$GCS_CERT"   # update for your cert on disk
 
-	export MONITORING_GCS_CERT_KEYFILE="$GCS_KEY"     # update for your private key on disk
-	
-	# Below are to enable GCS config download
-	export MONITORING_GCS_NAMESPACE=GHPILOGS
-	export MONITORING_CONFIG_VERSION=1.3
-	export MONITORING_USE_GENEVA_CONFIG_SERVICE=true
-        export MONITORING_TENANT=$tenant
-        export MONITORING_ROLE=GHPI
-        export MONITORING_ROLE_INSTANCE=${tenant}_1
+    export MONITORING_GCS_CERT_KEYFILE="$GCS_KEY"     # update for your private key on disk
+    
+    # Below are to enable GCS config download
+    export MONITORING_GCS_NAMESPACE=GHPILOGS
+    export MONITORING_CONFIG_VERSION=1.3
+    export MONITORING_USE_GENEVA_CONFIG_SERVICE=true
+    export MONITORING_TENANT=$tenant
+    export MONITORING_ROLE=GHPI
+    export MONITORING_ROLE_INSTANCE=${tenant}_1
 EOT
 
 ## Run container using Monitoring image, if not running already. Copy above created env variable files to container and start the cron job on running container..
@@ -133,16 +133,19 @@ MyContainerId="$(sudo docker run -it --privileged --rm -d --network host --name 
   fi
 
   echo -e "\nMonitoring container with Id $MyContainerId has started successfully...\n"
-    echo -e "Creating $GCS_CERT_FOLDER in the monitoring container"	
-    sudo docker exec -itd $MyContainerId bash -c test -d "$GCS_CERT_FOLDER" && sudo rm -f "$GCS_CERT_FOLDER/*" || sudo mkdir "$GCS_CERT_FOLDER" 
     
-    echo -e "Copying cert and key to the monitoring container"
-    sudo docker cp "$GCS_CERT" $MyContainerId:"$GCS_CERT" 	
-    sudo docker cp "$GCS_KEY" $MyContainerId:"$GCS_KEY"
-	
+    if [ -f "$GCS_CERT_WITH_KEY" ]; then
+      echo -e "Creating $GCS_CERT_FOLDER in the monitoring container"   
+      sudo docker exec -itd $MyContainerId bash -c test -d "$GCS_CERT_FOLDER" && sudo rm -f "$GCS_CERT_FOLDER/*" || sudo mkdir "$GCS_CERT_FOLDER" 
+    
+      echo -e "Copying cert and key to the monitoring container"
+      sudo docker cp "$GCS_CERT" $MyContainerId:"$GCS_CERT"     
+      sudo docker cp "$GCS_KEY" $MyContainerId:"$GCS_KEY"
+    fi
+    
     sudo docker cp ~/collectd $MyContainerId:/etc/default/collectd
     sudo docker cp ~/mdsd $MyContainerId:/etc/default/mdsd
     sudo docker exec -itd $MyContainerId bash -c '/etc/init.d/cron start'
-	
+    
  echo -e "Setting up of Monitoring container is successful.\n"
 fi
